The car can now drive in the simulator and stop at the first red light
The car can now drive in the simulator and stop at the first red light, but can't accelerate after the light turn to green. Modified the subscriber code line, and also the waypoint_cb function, so that it can publish the nearest red light stop line waypoints to the  vehicle/traffic_waypoint topic, and this message can be used by Waypoint_updater node to decelerate the car to stop at stop line.